+++ CHANGELOG.rst
@@ -19,7 +19,8 @@ Added:
 * ``ascatngs`` as a submodule
 * GitHub action to build and test ``ascatngs`` container
 * Reference section to docs/FAQ.rst
-
+* ascat download references
+* Delly tumor only rule
 
 Changed:
 ^^^^^^^^
@@ -31,6 +32,8 @@ Changed:
 * Update GNOMAD URL
 * Split tumor-only ``cnvkit batch`` into individual commands
 * Updated Snakemake to v6.5.3 
+* Improved TMB calculation issue #51
+
 
 Fixed:
 ^^^^^^
@@ -48,6 +51,14 @@ Removed:
 * Cleaned up unused container definitions and conda environment files
 * Remove cnvkit calling for WGS cases
 * Removed the install.sh script
+
+[7.2.5]
+-------
+
+Changed:
+^^^^^^^^
+
+* Updated COSMIC path to use version 94
 
 [7.2.5]
 -------